DevTools: didRecalculateStyleForElement should only be called if style really was recalced
Currently we call didRecalculateStyleForElement as we walk down the tree to find
the elements that actually need a style recalc and we don't ever call it when
doing a reattach of the element.

This moves the call into the if statement in Element::recalcStyle so we'll only
do it when actually recomputing the style or when doing an attach from inside
recalcStyle (which handles lazyAttach).

The count of the number of elements that went through a "Recalculate Style"
in the web inspector should be much more accurate.

I also updated timeline-recalculate-styles.html to exercise this new behavior. The
expected result is the same, but the number of nodes we walk through is much
higher.

TEST=LayoutTests/inspector/timeline/timeline-recalculate-styles.html

Review URL: https://chromiumcodereview.appspot.com/14842007

git-svn-id: svn://svn.chromium.org/blink/trunk@152854 bbb929c8-8fbe-4397-9dbb-9b2b20218538