Speed up png decoding by eliminating branches from the core loop that fills ImageFrame with pixels
The original idea came from the work of Allan Jensen landed in WebKit as r150252.
He eliminated the branches in PNGImageDecoder::rowAvailable to make it benefit from
auto vectorization. I tweaked it further to make this change a performance win
without auto vectorization as well (which we do not use by default, only with -O3 or
-ftree-vectorize). I measured the performance gain with callgrind by comparing the
inclusive cost of the affected function. I used simple local pages with a few image
for benchmarking.
Results:
  * opaque PNG: 7.8% faster
  * PNG with alpha: 24% faster

R=thakis@chromium.org

Review URL: https://chromiumcodereview.appspot.com/17448010

git-svn-id: svn://svn.chromium.org/blink/trunk@153004 bbb929c8-8fbe-4397-9dbb-9b2b20218538