RuleSet should use malloc rather than Vector
The lion's share of the memory used by the style resolver is in the RuleSet
objects. Prior to this CL, these objects were structured as a HashMap from
AtomicStrings to pointers to vectors of RuleData. This CL simplifies this
object graph by removing a layer of indirection. Now we just have a HashMap
from AtomicStrings to an array of RuleDatas.

Rather than use a length to terminate the iteration over the Vector, this CL
uses a bit in RuleData to mark the end of the array. Together with removing the
extra pointer, this CL saves 15 kB on Mobile Gmail.

R=eseidel,ojan

Review URL: https://chromiumcodereview.appspot.com/17616009

git-svn-id: svn://svn.chromium.org/blink/trunk@153042 bbb929c8-8fbe-4397-9dbb-9b2b20218538