layout tests: Print out the number of expected failures in the summary.
When running layout tests that are expected to fail, the output 'All X tests ran as expected.'
is not sufficiently clear that some of the tests failed, and are in fact expected to fail. So
include the number of tests that were expected to fail in the summary line.

BUG=none
R=dpranke@chromium.org

Review URL: https://codereview.chromium.org/18557002

git-svn-id: svn://svn.chromium.org/blink/trunk@153466 bbb929c8-8fbe-4397-9dbb-9b2b20218538