Add the NeedsRebaseline expectation value for TestExpectations.
This makes it more clear when people are adding lines just for the purposes
of rebaselining (e.g. [ImageOnlyFailure Failure Pass] could be a line added
for rebaselining or for a flaky test).

Also allows for creating a bot that automatically does the rebaseline once
the bots have cycled.

BUG=241364

Review URL: https://chromiumcodereview.appspot.com/18560005

git-svn-id: svn://svn.chromium.org/blink/trunk@153470 bbb929c8-8fbe-4397-9dbb-9b2b20218538