Add ImageFrame:setRGBARaw and use it for JPEG decoding
Add ImageFrame::setRGBARaw() and use it in the JPEG image decoder
for the very rare CMYK and RGB JPEG images cases to remove the code
branching caused by ImageFrame::setRGBA() in the row write loop.

Covered by existing tests, in particular:
  RGB jpeg: fast/images/rgb-jpeg-with-adobe-marker-only.html
  CMYK jpeg: fast/images/cmyk-jpeg-with-color-profile.html

The patch may improve JPEG decoding perf on Android, so we will be
observing the perf bots to find out.  No perf change is expected for
desktop chrome since the vast majority of web JPEG are YCbCr, and
they are decoded on a different code path.

Review URL: https://chromiumcodereview.appspot.com/18301011

git-svn-id: svn://svn.chromium.org/blink/trunk@153572 bbb929c8-8fbe-4397-9dbb-9b2b20218538