[CSS Grid Layout] Support named grid lines with 'span'
This change implements the resolution for named grid lines with 'span',
which is the missing piece to fully support named grid lines.

To resolve these cases, we need to:
- find the index of the first named grid line in the named grid lines
array of a certain name that is before or after the opposite position
(the direction depends on which property we are resolving).
- offset this index by the span position's <integer> to find the
right named grid line.

The specification doesn't detail this but it is strongly suggested by
the examples and the normal use cases (we don't want authors to have
to disambiguate the named grid lines manually if we can find a
sensible value).

BUG=234192
TEST=fast/css-grid-layout/grid-item-named-grid-line-resolution.html

Review URL: https://chromiumcodereview.appspot.com/17482006

git-svn-id: svn://svn.chromium.org/blink/trunk@153794 bbb929c8-8fbe-4397-9dbb-9b2b20218538