REGRESSION(153072): XMLHttpRequest#send incorrectly encodes 8 bit strings
When we added a special case for 8 bit strings, we mistakenly assumed that they
were already encoded in UTF-8 when in fact they're encoded in Latin-1. This CL
restore the correct behavior of always encoding the strings into UTF-8,
regardless of their internal representation.

R=eseidel

Review URL: https://chromiumcodereview.appspot.com/18966003

git-svn-id: svn://svn.chromium.org/blink/trunk@153841 bbb929c8-8fbe-4397-9dbb-9b2b20218538