Mixed content iframe checks should use top
Prior to this CL, we were using the parent frame when checking iframes for
mixed content. Unfortunately, that breaks the iTunes download page because it
uses an HTTPS iframe to kick off the HTTP download.

This CL causes us to use the top-level frame for this check. That's reasonable
because if the top-level frame is HTTP, the user doesn't have any expectation
that they're being protected from network attackers. This change restores our
behavior in this regard to match our behavior prior to
https://src.chromium.org/viewvc/blink?revision=150260&view=revision. We still
block mixed content iframes, which is the main change from
https://src.chromium.org/viewvc/blink?revision=150260&view=revision.

R=tsepez@chromium.org,cevans@chromium.org
BUG=244210

Review URL: https://chromiumcodereview.appspot.com/18104009

git-svn-id: svn://svn.chromium.org/blink/trunk@153879 bbb929c8-8fbe-4397-9dbb-9b2b20218538